requirements.txt
# Core Framework
Flask==2.3.3
python-dotenv==1.0.0

# WhatsApp & Communication
requests==2.31.0
twilio==8.9.0

# Database
supabase==1.2.0

# AI & NLP
anthropic==0.39.0

# Scheduling
APScheduler==3.10.4

# Utilities
pytz==2023.3
python-dateutil==2.8.2
icalendar==5.0.10

# Logging & Monitoring
colorlog==6.7.0

# Payment Processing
# Note: PayFast doesn't have an official Python SDK
# Using requests for API calls

# Data Processing
# Updated for Python 3.13 compatibility
pandas==2.2.3
numpy==2.3.3

# Security
cryptography==41.0.4

# Analytics
user-agents==2.2.0

# Voice Processing
pydub==0.25.1
# speech_recognition==3.10.0  # Optional, not currently used

# Testing (dev only)
pytest==7.4.2
pytest-cov==4.1.0

# Production Server
gunicorn==21.2.0

# Report Lab
reportlab==4.0.7
